Create apriltags.rs as another template
This file is untested, but I decided to upload it to make it clear what I'm trying to do, to make integration easier. This file does more processing, moving on from the baseline example. It will try to capture a frame from the camera, then try to detect apriltags. <Not functional and not integrated!>