Updating Prototype file networking.rs
Implemented separate non-blocking tasks that attempts to connect to each bot IP binded on separate ports, and a localhost connection as a debug test. The program will correctly open connections and listen on ports. Issues: Won't write to stream. Seems to be a deadlock or listener issue? Also, not error tested in other situations.

HMAC functionality coming after these issues are solved.